grocify_scraper/requirements.txt
# PDF Processing
pdfminer.six>=20221105
PyPDF2>=3.0.0
pypdf>=3.0.0

# HTTP & Downloads
requests>=2.31.0
urllib3>=2.0.0

# Data Processing
python-dateutil>=2.8.2

# JSON & Validation
jsonschema>=4.19.0

# Utilities
python-slugify>=8.0.0

# Optional: OCR (if needed)
# pytesseract>=0.3.10
# pdf2image>=1.16.3
# Pillow>=10.0.0

